Issue #692: Add SLF4J bridge and refactor logging and tool callback configurations
- Added `jul-to-slf4j` dependency to `pom.xml`
- Removed synchronized `update` method in `AgentContextHolder`
- Refactored `RestApplication` to use `setAgentContext` method
- Introduced `LoggingBridgeConfiguration` for JUL to SLF4J bridge
- Moved tool callback configuration to `ToolCallbackConfiguration`
- Updated logging levels in `application.yaml` and `logging.properties`
