+++ metricshub-agent/pom.xml
@@ -226,6 +226,11 @@
 			<artifactId>slf4j-api</artifactId>
 		</dependency>
 
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>jul-to-slf4j</artifactId>
+		</dependency>
+
 		<!-- CLI jansi -->
 		<dependency>
 			<groupId>org.fusesource.jansi</groupId>
