+++ metricshub-agent/src/main/resources/logging.properties
@@ -1,0 +1,4 @@
+handlers= java.util.logging.ConsoleHandler
+.level= WARNING
+org.apache.coyote.level = OFF
+org.apache.catalina.level = OFF
