+++ pom.xml
@@ -191,6 +191,11 @@
 				<version>2.0.11</version>
 			</dependency>
 			<dependency>
+				<groupId>org.slf4j</groupId>
+				<artifactId>jul-to-slf4j</artifactId>
+				<version>2.0.11</version>
+			</dependency>
+			<dependency>
 				<groupId>org.metricshub</groupId>
 				<artifactId>community-connectors</artifactId>
 				<version>${community-connectors.version}</version>
